--- lib/presentation/pages/repo/components/repo_sort_button.dart
@@ -1,25 +1,0 @@
-// Copyright 2022 susatthi All rights reserved.
-// Use of this source code is governed by a MIT license that can be
-// found in the LICENSE file.
-
-import 'package:flutter/material.dart';
-
-import '../../../../localizations/strings.g.dart';
-import 'search_repos_sort.dart';
-
-/// リポジトリ検索用ソートボタン
-class RepoSortButton extends StatelessWidget {
-  const RepoSortButton({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      onPressed: () => showModalBottomSheet<void>(
-        context: context,
-        builder: (context) => const SearchReposSortSelectorBottomSheet(),
-      ),
-      icon: const Icon(Icons.sort),
-      tooltip: i18n.sort,
-    );
-  }
-}
